feat: main is divided into modules
